Style about page
Used flex to seperate the two contents. On image on the about page, I set the width and height to 100 percent of the containter. I also used object fit cover to prevent the picture from stretching. I then used overflow hidden on its parent containter to prevent it from extending and preventing to scroll